Update dtw_variants.py
Some times you have ts with a nan in only one dimension. Ex: X=[[1,2],[1,nan]].  to_time_series_dataset wouldn't remove that time stamp. But computing the distance between X and any other ts, you would get a nan because on that dimension the diff goes to nan. The fix essentially make you skip the list computation only on that time stamp and that dimension that has a nan.