Update barycenter with nan computation
Given we compute the average of the ts values for each timestamp (following the assignement from dtw: (1/D)(sum_w_x) 
 
Where D is a vector (size of baricenter n_timestamps), with the number of the assigned ts for that timestamp of the centroid.
sum_w_x is the sum of the ts values assigned to that centroid timestamp. With dimension: (barycenter_size, d)
 We need make D as a matrix for each dimension, intead of one for each dimension and remove  for every timestamp and  dimension i where there are k np.nans in the ts for which sum_w_x was computing replacing nans with 0., with a matrix x_nan. So that for every dimension i we will compute D_adj_i = D-k_i
  Where k_i is the vector with the number of missing values on the ts, assigned to the timestamps, for that dimension.
 
So diag_sum_w_na = (1/D_adj_i)*sum_w_x_i. * is pairwise product